requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
flet
mysql-connector-python
plotly
pyrebase4
gTTS
flet-audio